Added a DefaultExporter for AI Tracing (#7085)
## Description

Added a DefaultExporter for AI Tracing. The exporter can run in one of 3
modes, as defined by the user and the storage provider:

- realtime
- batch-with-updates (for SQL-like DBs)
- insert-only (for high-volume prod DBs)

Currently the user must pass this exporter in the mastra Observability
config to use it. This will be fixed in another PR.

I also included the plan I made with Claude Code in the PR for reference
on the implementation.

## Related Issue(s)

#6773

## Type of Change

- [ ] Bug fix (non-breaking change that fixes an issue)
- [X] New feature (non-breaking change that adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to change)
- [ ] Documentation update
- [ ] Code refactoring
- [ ] Performance improvement
- [ ] Test update

## Checklist

- [ ] I have made corresponding changes to the documentation (if
applicable)
- [x] I have added tests that prove my fix is effective or that my
feature works

Will make documentation updates in another PR.