fix: mcp server in WSL installation (#7349)
## Description

This PR fixes the mcp server and ensures it works on `WSL`. 

The major update is fixing the log level so it aligns closely with the
spec and also adding a handler. The mcp inspector pings the server with
the ff message

```
{
  "jsonrpc": "2.0",
  "id": 1,
  "method": "logging/setLevel",
  "params": {
    "level": "info"
  }
}
```

and we weren't handling that, which led to the errors specified #7154. 

[MCP spec on
logging](https://modelcontextprotocol.io/specification/2025-06-18/server/utilities/logging)

## Related Issue(s)

<!-- Link to the issue(s) this PR addresses, using hashtag notation:
#123 -->

## Type of Change

- [x] Bug fix (non-breaking change that fixes an issue)
- [ ] New feature (non-breaking change that adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to change)
- [ ] Documentation update
- [ ] Code refactoring
- [ ] Performance improvement
- [ ] Test update

## Checklist

- [ ] I have made corresponding changes to the documentation (if
applicable)
- [ ] I have added tests that prove my fix is effective or that my
feature works