feat: authorize url in multi modal (#7244)
## Description

Adds a way to pass public URLs to the LLM for large documents when using
Multi Modal.

Note: since assistant-ui forces us to rely on a `File`object (with a
valid contentType matching what assistant-ui suppports) in the browser,
I had to implement a server endpoint to verify for the content type of a
given doc (cannot do that on the frontend because of CORS).



https://github.com/user-attachments/assets/0f62b0e2-6eb5-47a1-ada1-5aad980ed8c1


## Related Issue(s)

<!-- Link to the issue(s) this PR addresses, using hashtag notation:
#123 -->

## Type of Change

- [ ] Bug fix (non-breaking change that fixes an issue)
- [x] New feature (non-breaking change that adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to change)
- [ ] Documentation update
- [ ] Code refactoring
- [ ] Performance improvement
- [ ] Test update

## Checklist

- [ ] I have made corresponding changes to the documentation (if
applicable)
- [ ] I have added tests that prove my fix is effective or that my
feature works